BUGFIX: HTTPD server not to fail on missing dirs
Previously, if a request was issued for a missing path or file, the HTTP server would return a 404 which could result in deployment failures due to the deployment scripts becoming confused with the error status.

This change alleviates that making the HTTP server only fail with a 404 when attempting to fetch a non-existent *FILE*. If an attempt is made to fetch a non-existent directory, an empty directory entry will be returned.